+++ api/v1alpha1/httpchaos_types.go
@@ -73,6 +73,10 @@ type HTTPChaosSpec struct {
 	// Duration represents the duration of the chaos action.
 	// +optional
 	Duration *string `json:"duration,omitempty" webhook:"Duration"`
+
+	// RemoteCluster represents the remote cluster where the chaos will be deployed
+	// +optional
+	RemoteCluster string `json:"remoteCluster,omitempty"`
 }
 
 type HTTPChaosStatus struct {