+++ api/v1alpha1/timechaos_types.go
@@ -56,6 +56,10 @@ type TimeChaosSpec struct {
 
 	// Duration represents the duration of the chaos action
 	Duration *string `json:"duration,omitempty"`
+
+	// RemoteCluster represents the remote cluster where the chaos will be deployed
+	// +optional
+	RemoteCluster string `json:"remoteCluster,omitempty"`
 }
 
 // TimeChaosStatus defines the observed state of TimeChaos